requirements.txt
boto==2.49.0
boto3==1.10.26
botocore==1.13.26
dj-database-url==0.5.0
Django==2.2.5
django-storages==1.8
djangorestframework==3.10.3
docutils==0.15.2
gunicorn==20.0.2
jmespath==0.9.4
psycopg2==2.8.4
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
